+++ CHANGES.rst
@@ -16,7 +16,9 @@ Changes:
 
 Fixes:
 ------
-- No change.
+- Fix default `XML` format resolution for `WPS` endpoint when no ``Accept`` header or ``format``/``f`` query parameter
+  is provided and that the request is submitted from a Web Browser, which involves additional control logic to select
+  the applicable ``Content-Type`` for the response.
 
 .. _changes_4.36.0: